feat($1578985020000): remove history empty reusable view to change to just view
